Fixes
This commit fixes files overwritten by another commit:

ecwolf.svg was replaced with a low quality autotrace. Replaced it with my file from PR#26
fmtowns.svg was replaced with a logo with rendering problems--fixed
fmtownsmarty.svg was added, which is the original fmtowns.svg before i recreated it in my pull request #26. Replaced it with file from #26